+++ filter-service/src/main/java/ru/models/dto/Response.java
@@ -19,9 +19,9 @@ public sealed interface Response {
             List<String> market,
             List<String> blackList,
             String action,
-            int timeWindow,
+            long timeWindow,
             Direction direction,
             int percent,
-            int volume24h
+            long volume24h
     ) implements Response {}
 }
